Add files via upload
HSC 2023 Software Design and Development major work. This program takes in a list of numbers and the user can decide whether they want the list to be in ascending or descending order using a sorting algorithm (selection, insertion, and bubble sort). The program can also search whether a list has a certain element present in it. The program can also have a text-to-speech and black-and-white function allowing people with visual impairments to access the application easily.